requirements.txt
pyaudio
numpy
python-dotenv
deepgram-sdk
groq
cartesia
websockets
